# Swagger/swagger.json
{
	"components": {
...
		"schemas": {
			"DataResponse": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"data": {},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponseModel": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							"false"
						],
						"nullable": false
					}
				},
				"required": [
					"message",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AuthResponse": {
				"properties": {
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"token": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"statusCode",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SmsRequest": {
				"properties": {
					"phoneNumber": {
						"type": "string"
					}
				},
				"required": [
					"phoneNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SmsCodeRequest": {
				"properties": {
					"phoneCode": {
						"type": "string"
					}
				},
				"required": [
					"phoneCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Username": {
				"properties": {
					"username": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://localhost:3001/auth/google",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "Shopping Mall backend",
		"version": "1.0.0",
		"description": "shopping mall api documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/auth/register": {
			"get": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DataResponse"
								}
							}
						}
					},
					"201": {
						"description": "User Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseModel"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "method",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/Login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"403": {
						"description": "Invalid Password or Password",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseModel"
								}
							}
						}
					},
					"200 response": {
						"description": "Logged In",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						},
						"headers": {
							"DataResponse": {
								"schema": {
									"$ref": "#/components/schemas/DataResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SmsRequest"
							}
						}
					}
				}
			}
		},
		"/auth/update": {
			"post": {
				"operationId": "Update password",
				"responses": {
					"403": {
						"description": "Invalid Email",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseModel"
								}
							}
						}
					},
					"200 response": {
						"description": "Succesfully updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DataResponse"
								}
							}
						},
						"headers": {
							"DataResponse": {
								"schema": {
									"$ref": "#/components/schemas/DataResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SmsCodeRequest"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseModel"
								}
							}
						}
					},
					"201 response": {
						"description": "Logout",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DataResponse"
								}
							}
						},
						"headers": {
							"DataResponse": {
								"schema": {
									"$ref": "#/components/schemas/DataResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Username"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}